Fixed Timezone Date Warning - Fixed () error on aristotle server - Added another set of notifications to tell user they don't have classes listed

+++ edit_classes.php
@@ -109,34 +109,41 @@
 					<br>
 					<br>
 					<?php
-					if (mysqli_num_rows($r)){
 
-						echo '<table>';
-						echo '<tr>';
-						echo '<td><u>Subject</u></td>';
-						echo '<td><u>Code</u></td>';
-						echo '<td><u>Section</u></td>';
-						echo '<td><u>Name</u></td>';
-						echo '<td><u>Schedule</u></td>';
-						echo '<td><u>Professor</u></td>';
-						echo '<td><u>Room</u></td>';
-						echo '<td><u></u></td>';
-						echo '<td><u></u></td>';
-						echo '</tr>';
-						while ($row = mysqli_fetch_array($r)) {
+					if (isset($r)){
+						if (mysqli_num_rows($r)){
+
+							echo '<table>';
 							echo '<tr>';
-							echo '<td>'.($row['subject']).'</td>';
-							echo '<td>'.($row['code']).'</td>';
-							echo '<td>'.($row['section']).'</td>';
-							echo '<td>'.($row['name']).'</td>';
-							echo '<td>'.($row['schedule']).'</td>';
-							echo '<td>'.($row['professor']).'</td>';
-							echo '<td>'.($row['room']).'</td>';
-							echo '<td><a href="edit_class.php?id='.$row['class_id'].'">Edit</a></td>';
-							echo '<td><a href="delete_class.php?id='.$row['class_id'].'">Delete</a></td>';
+							echo '<td><u>Subject</u></td>';
+							echo '<td><u>Code</u></td>';
+							echo '<td><u>Section</u></td>';
+							echo '<td><u>Name</u></td>';
+							echo '<td><u>Schedule</u></td>';
+							echo '<td><u>Professor</u></td>';
+							echo '<td><u>Room</u></td>';
+							echo '<td><u></u></td>';
+							echo '<td><u></u></td>';
 							echo '</tr>';
+							while ($row = mysqli_fetch_array($r)) {
+								echo '<tr>';
+								echo '<td>'.($row['subject']).'</td>';
+								echo '<td>'.($row['code']).'</td>';
+								echo '<td>'.($row['section']).'</td>';
+								echo '<td>'.($row['name']).'</td>';
+								echo '<td>'.($row['schedule']).'</td>';
+								echo '<td>'.($row['professor']).'</td>';
+								echo '<td>'.($row['room']).'</td>';
+								echo '<td><a href="edit_class.php?id='.$row['class_id'].'">Edit</a></td>';
+								echo '<td><a href="delete_class.php?id='.$row['class_id'].'">Delete</a></td>';
+								echo '</tr>';
+							}
+							echo '</table>';
+						} else {
+							echo '<div style ="color: red">';
+							echo 'There are no classes displayed for this subject.';
+							echo '</div>';
 						}
-						echo '</table>';
 					} else {
 						echo '<div style ="color: red">';
 						echo 'There are no classes displayed for this subject.';
